requirements.txt
Flask==3.0.3
torch==2.3.0
Pillow==10.3.0
torchvision==0.18.0
gunicorn==20.1.0
requests==2.31.0
Flask-BasicAuth==0.2.0